requirements.txt
django>=4.2,<5.0
openai-whisper>=20231117
scikit-learn>=1.3
opencv-python>=4.8
numpy>=1.24
Pillow>=10.0
